fix: provider > singleton
프로바이더 대신 싱글톤패턴으로 코드변경
